+++ tests/field.test.tsx
@@ -1,7 +1,10 @@
 import React from 'react';
 import Form, { Field } from '../src';
 import type { FormInstance } from '../src';
-import { render } from '@testing-library/react';
+import { act, fireEvent, render } from '@testing-library/react';
+import { Input } from './common/InfoField';
+import { getInput } from './common';
+import timeout from './common/timeout';
 
 describe('Form.Field', () => {
   it('field remount should trigger constructor again', () => {
@@ -36,5 +39,38 @@ describe('Form.Field', () => {
     // expect(instance.cancelRegisterFunc).toBeFalsy();
     // expect((wrapper.find('Field').instance() as any).cancelRegisterFunc).toBeTruthy();
     expect(formRef.getFieldsValue()).toEqual({ light: 'bamboo' });
+  });
+
+  // https://github.com/ant-design/ant-design/issues/51611
+  it('date type as change', async () => {
+    const onValuesChange = jest.fn();
+
+    const MockDateInput = (props: { onChange?: (val: Date) => void }) => (
+      <button
+        onClick={() => {
+          props.onChange?.(new Date());
+        }}
+      >
+        Mock
+      </button>
+    );
+
+    const { container } = render(
+      <Form onValuesChange={onValuesChange}>
+        <Field name="date">
+          <MockDateInput />
+        </Field>
+      </Form>,
+    );
+
+    // Trigger
+    for (let i = 0; i < 3; i += 1) {
+      fireEvent.click(container.querySelector('button'));
+      await act(async () => {
+        await timeout();
+      });
+      expect(onValuesChange).toHaveBeenCalled();
+      onValuesChange.mockReset();
+    }
   });
 });
